Initially converted to TypeScript (not tested in FSM)
I just went ahead and converted this sucker to TypeScript and fixed all
the immediate tslint errors. It isn't set as a module and doesn't use
the module versions of sub-modules, so there are complication problems
with duplications there.

Also, there's no chance in hell this will actually work. Testing much
needed.